# Conflicts:
#	app.py

+++ app.py
@@ -15,6 +15,9 @@ from PyQt5.QtWidgets import (
 from simulator import Ui_PicSimulator
 
 
+win = QMainWindow
+
+
 class Window(QMainWindow, Ui_PicSimulator):
 
     def __init__(self, parent=None):
@@ -34,9 +37,9 @@ class Window(QMainWindow, Ui_PicSimulator):
 
     def fLoadFile(self):
         simulationParser.queue = []
-        print(simulationParser.lst)
+        # print(simulationParser.lst)
         simulationParser.lst = []
-        print(simulationParser.lst)
+        # print(simulationParser.lst)
         simulationParser.get_file()
 
         for row in range(self.showCode.rowCount()):
@@ -53,14 +56,18 @@ class Window(QMainWindow, Ui_PicSimulator):
             item.setCheckState(QtCore.Qt.Unchecked)
 
             self.showCode.insertRow(i)
-            self.showCode.setItem(i, 0, item)
+
+            if not(str(simulationParser.lst[i][0]).strip() == ""):
+                self.showCode.setItem(i, 0, item)
+
             self.showCode.setItem(i, 1, QTableWidgetItem(str(simulationParser.lst[i][0])))
             self.showCode.setItem(i, 2, QTableWidgetItem(str(simulationParser.lst[i][1])))
             self.showCode.setItem(i, 3, QTableWidgetItem(str(simulationParser.lst[i][2])))
             self.showCode.setItem(i, 4, QTableWidgetItem(str(simulationParser.lst[i][3])))
             self.showCode.setItem(i, 5, QTableWidgetItem(str(simulationParser.lst[i][4])))
             self.showCode.setItem(i, 6, QTableWidgetItem(str(simulationParser.lst[i][5])))
-            # marks table row
+
+             # marks table row
 
     def fButtonStart(self):
 
@@ -88,12 +95,14 @@ class Window(QMainWindow, Ui_PicSimulator):
 class FindReplaceDialog(QDialog):
 
     def __init__(self, parent=None):
+
         super().__init__(parent)
 
         # loadUi("ui/find_replace.ui", self)
 
 
 if __name__ == "__main__":
+
     app = QApplication(sys.argv)
 
     win = Window()
